sbpy.activity.gas overhaul in order to include a new production rate module
