+++ sbpy/activity/gas/__init__.py
@@ -1,0 +1,3 @@
+from .. core import *
+from .core import *
+from .productionrate import *
