+++ sbpy/activity/gas/core.py
@@ -29,6 +29,7 @@ __all__ = [
 
     'Haser',
     'Vectorial',
+    'GasComa'
 ]
 
 from warnings import warn
@@ -46,10 +47,10 @@ except ImportError:
 
 from astropy.table import Table
 from astropy.utils.exceptions import AstropyWarning
-from .. import bib
-from .core import (Aperture, RectangularAperture, GaussianAperture,
+from ... import bib
+from .. core import (Aperture, RectangularAperture, GaussianAperture,
                    AnnularAperture, CircularAperture)
-from .core import rho_as_length
+from .. core import rho_as_length
 
 
 def photo_lengthscale(species, source=None):
